Fixed Kaggle T4 GPU BFloat16 hardware incompatibility

+++ generate_lora_kaggle.py
@@ -15,7 +15,7 @@ def generate_lora_kaggle():
     # 1. Compress and load the base brain
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True,
-        bnb_4bit_compute_dtype=torch.bfloat16
+        bnb_4bit_compute_dtype=torch.float16
     )
     
     tokenizer = AutoTokenizer.from_pretrained(model_id)
